pagination on search results

+++ web/src/main/java/com/itechart/web/service/files/FileService.java
@@ -1,6 +1,7 @@
 package com.itechart.web.service.files;
 
 import org.apache.commons.io.FileUtils;
+import org.apache.commons.lang3.StringUtils;
 
 import java.io.File;
 import java.io.IOException;
@@ -22,8 +23,12 @@ public class FileService implements AbstractFileService {
 
     public void deleteFile(String name) {
         // TODO: 29.03.2017 check null
-        File file = new File(FILE_PATH + FileSystems.getDefault().getSeparator() + name.charAt(0) + FileSystems.getDefault().getSeparator() + name);
-        if (file.exists()) {
+        if (StringUtils.isEmpty(name)) return;
+        String filePath = StringUtils.join(new Object[]{FILE_PATH, name.charAt(0), name}, FileSystems.getDefault().getSeparator());
+        File file = null;
+        if (filePath != null)
+            file = new File(filePath);
+        if (file != null && file.exists()) {
             try {
                 FileUtils.forceDelete(file);
             } catch (IOException e) {
@@ -41,14 +46,15 @@ public class FileService implements AbstractFileService {
     }
 
     public byte[] getFile(String name) {
-        if (!name.isEmpty()) {
+        if (StringUtils.isNotEmpty(name)) {
             // TODO: 30.03.2017 check file exists
-            Path path = Paths.get(FILE_PATH + FileSystems.getDefault().getSeparator() + name.charAt(0) + FileSystems.getDefault().getSeparator() + name);
-            try {
-                return Files.readAllBytes(path);
-
-            } catch (IOException e) {
-                e.printStackTrace();
+            Path path = Paths.get(StringUtils.join(new Object[]{FILE_PATH, name.charAt(0), name}, FileSystems.getDefault().getSeparator()));
+            if (path != null) {
+                try {
+                    return Files.readAllBytes(path);
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
             }
         }
         return null;
